refactor: generator 패키지 분리 및 상수 분리
