+++ src/main/java/lotto/controller/LottoController.java
@@ -5,8 +5,8 @@ import lotto.model.domain.LottoMoney;
 import lotto.model.domain.LottoNumber;
 import lotto.model.domain.LottoResult;
 import lotto.model.domain.Lottos;
-import lotto.model.domain.RandomNumbersGenerator;
 import lotto.model.domain.WinningNumbers;
+import lotto.model.domain.generator.RandomNumbersGenerator;
 import lotto.model.service.LottoService;
 import lotto.view.LottoInputView;
 import lotto.view.LottoOutputView;
