+++ src/main/java/lotto/model/domain/AutoLottos.java
@@ -2,10 +2,11 @@ package lotto.model.domain;
 
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.model.domain.generator.NumbersGenerator;
 
 public final class AutoLottos extends Lottos {
 
-    public static final int START_INCLUSIVE = 0;
+    private static final int START_INCLUSIVE = 0;
 
     public AutoLottos(final int size, final NumbersGenerator numberGenerator) {
         super(IntStream.range(START_INCLUSIVE, size)
