+++ src/main/java/lotto/model/domain/LottoNumber.java
@@ -4,8 +4,8 @@ import java.util.Objects;
 
 public final class LottoNumber {
 
-    public static final int MINIMUM_NUMBER = 1;
-    public static final int MAXIMUM_NUMBER = 45;
+    private static final int MINIMUM_NUMBER = 1;
+    private static final int MAXIMUM_NUMBER = 45;
 
     private final int value;
