+++ src/main/java/lotto/model/service/LottoService.java
@@ -4,9 +4,9 @@ import lotto.model.domain.AutoLottos;
 import lotto.model.domain.LottoMoney;
 import lotto.model.domain.LottoResult;
 import lotto.model.domain.Lottos;
-import lotto.model.domain.NumbersGenerator;
 import lotto.model.domain.RankResults;
 import lotto.model.domain.WinningNumbers;
+import lotto.model.domain.generator.NumbersGenerator;
 
 public final class LottoService {
