+++ src/test/java/lotto/model/domain/AutoLottosTest.java
@@ -3,6 +3,7 @@ package lotto.model.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.model.domain.generator.FixedNumbersGenerator;
 import org.junit.jupiter.api.Test;
 
 class AutoLottosTest {
