+++ src/test/java/lotto/model/domain/LottosTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.model.domain.generator.FixedNumbersGenerator;
+import lotto.model.domain.generator.NumbersGenerator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
